This commit fixes the HSD issue "Camera application crashes on Froyo when using zoom feature (2mp or 5mp camera)".
This is a workaround to
prevent camera apk crash when utilizing zoom functionality.
For optimizations we'd want to implement a solution here where
we set up zoom ratios for each of the supported platform cameras.
See the following files for more information on using the
KEY_MAX_ZOOM[] & KEY_ZOOM_RATIOS[] for finer grained control:
  frameworks/base/include/camera/CameraParameters.h
and
  frameworks/base/libs/camera/CameraParameters.cpp

NOTE: This commit only fixes the zoom crash issue, in testing with our CDK
and the 2mp and 5mp camera the zoom functionality does not allow the slider
to change zoom ratios (using Google AOSP defaults). I don't know if our
platform camera's driver/libcamera.so need extra functionality added to enable
zoom. This would need to be investigated further, but I don't know if that would
be myself or the camera driver team.

HSD=3452792

Change-Id: I1156e309181f2e7d5c5b7bd2ea0b93f30bc9669e
Signed-off-by: Brian Wood <brian.j.wood@intel.com>