camera_hal: Fixed video and preview switch issue
[BZ486 and BZ720]
When switching from preview to video, or switch the video recording resolutions,
Got DQ errors from driver sometime and the application has no response any more.

Caused by 3A thread which is not stopped when we do the stopPreview.
Add a wait to wait for 3A to stop in stopPreview.

Change-Id: I05e938a3700f351eb43aa210885c5333c1377e80
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Signed-off-by: Xiaolin Zhang <xiaolin.zhang@intel.com>