libcamera: remove obsolete exposure controls
Borqs-Id: 582d2be247a138184fbddb38772c10eba89c336c
Borqs-Contact: Jozef Kruger <jozef.kruger@intel.com>
Commit-Date: Tue Aug  2 03:12:33 PDT 2011

BZ: 5688
This is part of the changes in this BZ. We defined a new control to
transfer exposure settings to the sensor drivers. All other exposure
controls are now obsolete. The camera HAL used some of them that are
now obsolete, so we removed them.

Change-Id: Ib0704892c0ee44335d96620a7e1348b1d8f0ae97
Gerrit-Id: Ia858be5a91fab091ec6a5d8979003bee365de676