libcamera: unify focus driver interface
This patch unifies the auto-focus part of the sensor driver interface
across both iCDK and PR2. This allows us to remove work-around code
from the libmfldadvci library.

This patch also caches the previous AWB output values and makes sure that
whenever the new values are the same as the old AWB values no further
recalculations are done. This should save some compute power on the host.

This patch also moves the sensor_mode_data struct from atomisp_platform.h
to the PR2 sensor driver (mt9e013.c). This structure is highly sensor
specific, so it should not appear outside of the driver (and the 3A tuning
parameter file). I made atomisp_platform.h include atomisp.h for the
atomisp_sensor_mode_data struct which caused a compiler error due to a double
declaration of a BC_Video struct. I resolved this by including the header
file that contains the original declaration of this struct in atomisp_ioctl.c
and let atomisp.h export a renamed version of this struct
(atomisp_bc_video_package). I also moved the illegal external declaration
of the BC_Video function to the rightful owner (bufferclass_video_linux.h).
This means that the function prototype is now checked by the compiler against
its implementation and we no longer have a duplicate structure type definition.

Change-Id: I9375505d8d4cbca362f55e80ce8a4205a0df40d4
Signed-off-by: Jozef Kruger <jozef.kruger@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Signed-off-by: Xiaohai Xu <xiaohai.xu@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>