libcamera: Don't stop preview when preview is not running
Root cause:
Regression by patch 15795. When the preview is not running, the
stopPreview run anyway. So in burst capture, the streamoff called
before the DQ. DQ error caused the ISP reset. This cause the camera hang.

Solution:
Check the preview status before stop the preview.
Also add error when startSnapshot failed in burstcapture.

Change-Id: Ib620dd4627206a8116f9f8e1c8c12582d6681728
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Signed-off-by: Xiaohai Xu <xiaohai.xu@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>