flip ISP image vertically
BZ : 12950

When Taking a picture with world facing camera, it shows mirror image

Rootcause:
The image output from ISP was mirrored vertically, so it needs to flip
vertically to avoid mirror image when taking a picture.

This also fixes recorded video mirrored bug.

The image was fliped in cameraserver when preview, so camera preview
doesn't have this problem.

Modify attribute.orientation to correct picture display orientation.
Because now mt9e013 is unavailable, it will go to else in setupPlatformType(),
so mSnapshotSkipFrame needs to set as 2 to avoid dark picture.

Depends-Change-Id: Ib4f09b31375f9f363abc926753e98cd997819fbb
Depends-Change-Id: I7b08d2b770c5e623467278f8df4b07ee42aefc5f
Depends-Change-Id: Icd69d3834f78be1c3a3b1b89056b09df418ead21
Change-Id: Ib61955211cda4da15423aca2b5ffe05cd97d027d
Signed-off-by: Xiaohai.Xu <Xiaohai.Xu@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>