libcamera: Camera HAL redesign
This patch consists of the skeleton code for the camera hal design.
HAL will consist of four threads: Control, Display, JPEG, and AAA.

At this point, the data flow is working with a stubbed out
driver and a basic test app. State machine can be started and
stopped, and data can flow from thread to thread without any
known race conditions or crashes.

This is still a work in progress. There are a number of items that
need to be done that are marked with comments tagged "TODO".

Skeleton needs to be filled out and hooked up to the driver and
libmfldavci library.

Change-Id: I0a247df4ad80a4b78cdf28cc72cff03872aed234
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>