libcamera: Validate buffers against current state
When the camera HAL transitions between states, it's possible for
frames to be queued to the driver in the wrong state. For example,
if in preview mode and we try to take a picture, a preview frame may
get queued to the driver when it is in snapshot mode.

Issue: ANDROID-1333
Change-Id: Ie67b17b55f8ccec3e98078ed382972f07341a7ce
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>