libcamera: restart isp when video format changes
The ISP needs to be restarted when the video format changes in video mode.
AtomISP class should not be starting and stopping the ISP without the
ControlThread knowing about it. Calls that start/stop the ISP result in
buffers being deallocated. If buffers are floating around in the camera
stack when they are deallocatd, system will result in undefined behavior.

Issue: ANDROID-1335
Change-Id: I70775a4fae0fd6dc596c9888b4e3d40634e45251
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>