libcamera: Clean up parameters
1. Parameters are grouped into two different types:
Static parameters: Parameters that need to be set before preview or
recording is started. Typically, static parameters will have on affect the
format or size of camera buffers. If static parameters are set while camera
is running then the ISP needs to be stopped, threads need to stop, ISP needs
to be reconfigured, buffers need to be deallocated and reallocated, ISP
needs to be restarted, threads need to be restarted.

Dynamic parameters: Parameters that affect features during runtime. Dynamic
parameters include focus, zoom, and others. Setting dynamic parameters shouldn't
require the ISP to be restarted.

2. Default configuration moved to AtomISP class:
Default configuration is mostly dependent upon hardware and driver settings.
The closer we keep the default coniguration to the hardware, the more flexible
the design will be.

3. Parameter organization:
Move all parameters, types, constants that should be private to the AtomISP class
from the AtomCommon.h header to AtomISP.cpp. This will prevent other classes from
using types they shouldn't be using, encouraging information hiding and preventing
coupling of classes and modules. The goal is to help keep the design modular and
flexible.

4. Start and stop preview helper methods:
Implement helper methods so ISP can be restarted easily if static parameters are
changed while preview is running.

Issue: ANDROID-1398
Issue: ANDROID-1468
Change-Id: I89789ef254e91963c38905144505b76108b8b81e
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>