libcamera: single zoom ratio for video
This patch is an ammendment to the patch "Add support for zoom".
"Add support for zoom" seems to have exposed a limitation in the
default camera application where video preferences have a dependency
on zoom support. So if zoom is not supported, then the app does not
render camera preference controls on the UI for video mode.

This patch works around that issue by indicating that zoom is supported
in video mode, but only indicats a single supported zoom value (100).

Issue: ANDROID-394
Change-Id: I7f31c2d76cce5ad7cc13241f8c597daebef23504
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>