libcamera: Dynamic number of preview/record buffs
Previously buffers were hardcoded to 4. But for buffer sharing we
need to be able to support a dynamic number. All buffer storage
arrays are made from static to dynamically allocated.

Also, since buffer objects are dynamically allocated, we can no longer
pass pointers to AtomBuffer around the HAL. Instead, copies of the buffer
structure must be passed around to prevent access to freed memory. If copies
are not passed around, then freed memory is accessed when buffers are returned
to the control thread after it has already stopped the ISP (which is when it
deallocates its buffer arrays).

Change-Id: I034ccf7676e47c970441cfe5a36ce4af892454f1
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>