libcamera: Implement VideoThread
VideoThread issues dataTimestamp callbacks to the upper layers. VideoThread
was introduced so that ControlThread does not have to issue callbacks in order
to prevent deadlocks.

Deadlocks occur when the CameraService holds a lock while calling into the
ControlThread. Meanwhile the ControlThread is trying to obtain the same lock
in one of the callbacks. Since The ControThread can't obtain the lock, it
never returns from the callbacks, and is thus unable to respond to the
CameraService.

For: ANDROID-946
Change-Id: Iee2c62ffc0ffd95586f178a31d55e1016202b248
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>