libcamera: fix 3A lock modes
We used the supported modes for auto-focus, auto-whitebalance and
auto-exposure locks to publish the true,false modes we support.
But, the way these parameters are used by Android we should only set
to true if we support that lock or false if we don't support it.
By default, Android uses true and false to enable/disable one of 3A
lock.

Issue: ANDROID-1353
Issue: ANDROID-1354
Change-Id: I9c56327aa068906bf734f1873bdc97f5fd75bc82