libcamera: Implement shared buffers for capture
Add implementation in order to use the libva buffers from libjpeg
directly into AtomISP when in CAPTURE mode.
This way, AtomISP will use the snapshot buffers obtained from
libva in libjpeg.
Thus, when using libjpeg in order to compress the picture we will
just use the shared buffers instead of copying the data.

This is the stage 2 in burst-capture mode implementation. Although,
this patch does not implement yet the burst-capture, it adds a
performance improvement to still image captures: from 500ms to
300ms to encode an 8MP image.

Change-Id: Ie957d50af4a11dbc72051cc0c4ce8e09088f17cc