Fix invalid buffer causing FD crash
Face detection was crashing because it was holding stale pointer to
buffer data (wrapped in AtomBuffer objects) that is de-allocated when
we restartPreview because of setParameter calls.

The fix in this change is following:

(1) controlThread wait for FD to finish before stopISP which deallocates buffer
(2) controlThread also check AtomBuffer sessionID for validation before it sends to FD.

Basically these changes eliminate the chance for FD to hold/use any stale buffer pointers.

Because of change for (1), deadlock is observed between FD callback and setParameters.
The root cause of the deadlock is fixed in CameraService as dependency specified below.

Also because of change for (1), we added new API in IFaceDetector to allow its client to
specify whether to wait for FD to finish processing buffers.

Issue: ANDROID-1881

Change-Id: I4e66fd0ab561b14795cbc752b192962a14ac0c8e
Depends-Change-Id:I4e53d5c102d85b30cc741766a30500458a79c6f5