Removed video zoom workaround from camera HAL.
This change removes the workaround that enabled zoom with
one ratio in video mode to prevent the Camera app from
hiding the other setting controls due to a bug. The Camera
app has been fixed so it no longer hides the other
settings when zoom is not supported. This change sets
KEY_ZOOM_SUPPORTED to false when in video mode to prevent
the zoom control widget from being displayed in the Camera
app while in video camera mode.

Issue: ANDROID-394
Issue: ANDROID-1481

Change-Id: I1dfd07061abbab0706e9dd45a3811bea669fe968