libcamera: Fix dead-lock in AAAThread
The threadLoop implementation from AAAThread was returning status ==
NO_ERROR in case of exit.
So, when requestExitAndWait was called, mThreadRunning was set to false
and threadLoop should exit. But, as specified in Thread class, a true
return from threadLoop means that this method should be called again.
A false return from threadLoop won't call again this function.
Therefore, if (status == NO_ERROR) was true, threadLoop was called again
and we are blocked in waitForAndExecuteMessage().
Since there is no one to send a message after requestExitAndWait call, a
dead-lock resulted.

There is no issue opened for this bug, I found it when I was trying to
optimize the "camera on to first shot latency".

Change-Id: Ifdf23c605b2c7152827d2b9b3c81ad1e04910356