MUST_REVERT: workaround for libjpeg memory leak
When using libva shared buffers in libjpeg, these buffers are NOT
released if we are not calling jpeg_start_compress between
jpeg_create_compress and jpeg_destroy_compress.

So, if between these two calls we don't make any call to
jpeg_start_compress, then make a fake call so that libjpeg will
release the libva resources.

For: ANDROID-1998
Change-Id: If8783b2b5b6ed142a8350e619c590a44f93774ce