Have the preview frames be in NV21 format to the app
CDD requires the default format of preview frames delivered to an
application be in NV21 format.  This change maintains the use of
NV12 from the ISP to the preview surface, but converts the data to
NV21 when delivering to an application (IF the application requires
preview frame data).  It also correctly reports NV21 as the only
format we support for this frame delivery to the app.

Issue: ANDROID-1954
Change-Id: Ie7d1e6cc608841b0c6ff5c103906395816058113
Signed-off-by: ken lierman <ken.lierman@windriver.com>
Signed-off-by: Kwanhong Lee <kwanhong.lee@windriver.com>