libcamera: Fix auto-focus logic
Currently, after the autoFocus sequence our focus position remains
locked since the auto-focus was locked in auto-focus sequence and is
never unlocked after this.

When camera client is calling autoFocus in PREVIEW mode we are running
the auto-focus sequence in 3A library to get a sharp focus. Once the
auto-focus sequence is done (either fail or success, as libmfldadvci
library reports) we lock the auto-focus, so when takePicture will be
called we have the focus position established in auto-focus sequence.
But, if the camera client  won't call anymore the takePicture method,
it will call cancelAutoFocus. Here, we should unlock the focus and
switch back to auto-focus enabled in 3A.
Also, after takePicture, when camera client is calling startPreview,
we should unlock the auto-focus and switch back to auto-focus in 3A.

Change-Id: I87f97cbec64b26b3ac4dfe426ec25403d59655ac