libcamera: Fix video aspect ratio in preview video
When the camera client wants preview only but it also sets the recording
hint, we are starting ISP in video mode. But, since the camera client
wants only preview it will set only the preview size. The video size
will
be the default one, set in AtomISP::getDefaultParameters. The recording
hint does not mean that the application MUST send us the video size too.
It only means that we should be prepared for a startRecording. This is
the main reason for which we are starting ISP in VIDEO mode and not in
PREVIEW mode.
Since the default video size is 640x480 in AtomISP::getDefaultParameters
if the camera client wants a preview of size 640x360, we will start ISP
with 640x360 in viewfinder (preview) and 640x480 in video (recording).
But, the ISP is using the aspect ratio from video to get frames from
sensor, so it will use the same aspect ratio for viewfinder. But, since
the viewfinder size has a size with different aspect ratio, it will
stretch the image to fit into the requested size.

The fix for this situation is that we, camera HAL check the preview
aspect ratio and video aspect ratio every time we receive parameters and
make sure that preview and video have the same aspect ratio, if only the
preview size has been requested.

Issue: ANDROID-1632
Change-Id: I016fcacd6f62392e23108e31b02924338ad5419d