libcamera: Override preview resolution for video
Workaround: The camera firmware doesn't support preview dimensions that
are bigger than video dimensions. If a single preview dimension is larger
than the video dimension then the FW will downscale the preview resolution
to that of the video resolution.

This patch ensures that all HAL components are aware of the actual preview
resolution that comes from the camera firmware.

Issue: ANDROID-2157
Change-Id: I894688967a2faf6e040a9091e90012d390e972a0
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>