libcamera: Remove calls to MessageQueue::clearAll
All calls to clearAll have been replaced with remove to delete specific message
on an as needed basis. clearAll was initially introduced because we wished to
flush buffers. But because the MessageQueue class had a circular buffer
implementation it was impossible to remove specific messages.

This patch fixes two bugs:
- CallbacksThread::MESSAGE_ID_JPEG_DATA_READY was being removed and a memory
  leak resulted because the JPEG buffers weren't being deallocated.
- PreviewThread::MESSAGE_ID_SET_PREVIEW_WINDOW was being removed and as a result
  a vaild preview window was not propagating to the PreviewThread at the start
  of a new session. The PreviewThread continued to use the window from the previous
  session which resulted in a crash.

In each case where clearAll was replaced with remove the intention is for buffers
to be flushed at the end of a session so that the buffers can be deallocated safely.
We flush the buffers to prevent other threads from accessing invalid memory regions.

At this time, no other messages have been identified as needed to be removed. As opposed
to wiping a thread's messages, a thread is sent a message to manage its own state safely
in its own thread context.

Because of its risky nature the actual clearAll method will be removed in a subsequent patch.

For: ANDROID-1969
Change-Id: I231c6f6664fca17b58a8cbf45f19258124dad01b
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>