libcamera: Remove MessageQueue::clearAll method
Due to its risky nature, the clearAll method is removed to prevent all of
a thread's messages from being wiped out. Some messages dictate a thread's
state or trigger memory deallocations and must be processed by the thread.

If a thread message needs to be deleted the MessageQueue class has a remove
method for removing specific messages on an as needed basis. This is a much
safer approach.

For: ANDROID-1969
Change-Id: I0740de195f51c56d5a48b8e5648926505fbd6dfe
Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>