Switch preview mode when necessary
This change adds a check for which preview mode is active
when a still or video recording is requested and switches to
the correct mode as necessary.

This fixes the android.hardware.cts.CameraTest.testRecordingHint
CTS failure which verifies a still picture can be taken when
the hint indicates video recording, and that a video recording
can be taken when the hint indicates still preview mode.

Issue: ANDROID-2413
Change-Id: Ia86b321e7acf154fe05aae9ab1497cd1b412de6b
Signed-off-by: Joe Vitkay <joe.vitkay@windriver.com>