Conditionally rotate DV1 camera
As long as there is a different build for DV1 since it has
different firmware from the Plan-Of-Record DV2 platform, we
might as well have a conditional in the camera code that
rotates the camera upright if the platform is a DV1.

This patch propagates the build-environment TARGET_PLATFORM
variable to a #define c-flag and then does the appropriate
conditional-code thing to rotate the camera for DV1 to have
the same orientation as the rest of the system.

For: ANDROID-1043
Change-Id: I3bd371710961d79b81c0ec72d03d992840c505fc
Signed-off-by Paul Drews <paul.drews@intel.com>