libcamera:HDR: Add default HDR keys in the params structure
Problem: Test Camera Application crashes with a NULL pointer when
checking the HDR related keys in the camera parameters, because
they are not there.

Solution: add those keys to the camera parameters that the HAL fills
with defaults value indicating that HDR is not supported
Later on when we add support we can change this.

Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>