libcamera: Fix face detect callback hang
The HAL was hanging while stopping camera preview. The ControlThread
was waiting for the face detection thread to stop, but the face detection
thread was stuck trying to callback with face meta data because
CAMERA_MSG_PREVIEW_METADATA was disabled.

This patch moves the callback from the face detect thread to the
CallbacksThread which prevents the face detect thread from hanging
when CAMERA_MSG_PREVIEW_METADATA is disabled.

Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>