libcamera: Fix issue with ControlThread references
There appears to be an issue with the way that the Thread base class
handles reference counting. Thread will release its reference count
to the Thread object at the end of the Thread::_threadLoop() method.
However, requestExitAndWait will often return before _threadLoop
decrements its reference count. So if the caller of requestExitAndWait
calls clear, there is no guarantee that the thread object will be
destroyed at that time (since Thread base may still be holding a
reference).

In our case, we expect the Thread object to be destroyed before
ATOM_CloseCameraHardware so that when ATOM_OpenCameraHardware
is called, we don't have a lingering reference to the ControlThread
instance that was used in the previous camera session.

To work around this, this patch creates two new ControlThread methods:
- init()
- deinit()

All initialization was moved from the constructor to init(), and all
deinitialization was moved from the destructor to deinit(). This way,
it makes no difference if we create a new instance of ControlThread
while there is a lingering reference to the previous instance as long
as deinit() was called on the previous instance before init() is called
on the new instance.

Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>