libcamera: Implement Raw Callbacks
This patch implements two callbacks:
- The one for notification of Raw image
- The one to send the raw data back to the user.

In order to do so we have changed the flow how the raw frames
( postview and full-res snapshot YUV) are used.

Now when ControlThread finishes the takePicture process, it
sends both frames to two diferent Threads:
-  PictureThread for jpeg encode and to
-  CallbackThread to be given back to user (if requested)

The raw frames will wait in the CBThread until the JPEG encoding
process is completed. Once the CBThread receives the compressed
image it will return the raw frames back to the ControlThread.
It will use the ICallbackPicture that was used before
by the PictureThread. This now has been moved to the CBThread