libcamera: Flash feature updates
Limit flash modes to: auto, on, off, and torch. Remove
support for day sync and slow sync.

Update supported flash modes to correspond to scene mode
selections.

Fix torch mode so that flash is immediately turned on when
enabled and stays on until torch is disabled.

Simplify logic for flash state.

Signed-off-by: Robert Crabtree <robertx.l.crabtree@intel.com>
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>