Fix for testRecordingHint CTS test-case
The testRecordingHint CTS test-case was failing because it was using
startRecording takePicture calls combination with the recording-hint
parameter set on TRUE.
Because of some bug in the HW encoder, the HW JPEG encoder was hanging
when using buffers allocated before a recording session.
In order to fix the issue caused, a workaround was made in the code.