libcamera2: Fix for stopCapture in setParameters
BZ: 38158

When setParameters is called right after takePicture, there is a high
chance that the actual JPEG buffer to be deleted before sent to camera
client.
And this is because the flushPictures method from CallbacksThread is
removing MESSAGE_ID_JPEG_DATA_REQUEST and MESSAGE_ID_JPEG_DATA_READY
messages.
Those type of messages should be processed before processing the
MESSAGE_ID_FLUSH, because if there is a JPEG request (which means that
the camera client is waiting for a COMPRESSED callback) and we remove
and clear all the JPEG buffers, then we will have a hang (camera client
is waiting for COMPRESSED callback which will never be called by the
HAL, since we just removed the JPEG buffers.

Therefore, the normal way would be to let the CallbacksThread process
all the JPEG buffers and let it send the callbacks (if data was
requested) or just collect the JPEG buffers and release them all when
MESSAGE_ID_FLUSH is being processed.

Change-Id: Ieb46b74d56000de1bdd75f28a1292b059a107fb9
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/51437
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>