libcamera2: Fix for setParameter call handling while capturing
BZ: 38158

When setParameters is called during capture (which is the state right
before the first takePicture call), we need to check for the new picture
size (since this can affect the capture in progress).
If the picture size has changed we need to stop the capture.
In order to avoid dead-locks, the stopCapture needs to be done outside
setParameters handling method.

Change-Id: Iadf6351d4b7b7c0dc98d9598ccb4e6f9278cfff3
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/51439
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>