libcamera2: [PORT FROM ICS] Add fields to AtomBuffer to describe buffer type
BZ: 42181

Extensions to AtomBuffer struct needed to handle the zerocopy preview
data flow.

New type enum is added to better serve the use, as mode enum is more of
an ISP mode, not buffer. With this type we can distinguish between
preview allocated by AtomISP and preview buffers dequeued from a native
window. Extra fields are added to store relevant information for this
use

Re-formatted comments in code.

Change-Id: Ie1c3c45be8afbf13bd1505bee0c245e4dc944c35
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/52958
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/57643