libcamera2: AtomISP: clean up camera id handling
BZ: 38158

In AtomISP, there are two integer indixes related to camera input/sensors.
One is the Android 'cameraId' that upper layers use to select which camera
to use. App layer uses AtomISP::getNumberOfCameras and AtomISP::getCameraInfo()
to query the available cameras. Once it has picked which camera to use,
this id is passed to AtomISP() constructor.

Another index is the V4L2 input enumeration index, which is used with
VIDIOC_ENUMINPUT ioctl. While this index refers to a camera input/sensor source
as well, this is not equivalent to the Android camera.h cameraIds.

This patch cleans up how these two indexes are used. The intent is to
make the disctinction between the two indexes clear. The static index tables
are also renamed to 'sCamInfo' and 'sCameraInfo' to follow the Android
coding guideline for static member variables.

Change-Id: I05d7d68f1d0b497f3afde0032e503d18c1e69eb2
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/52004
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>