libcamera2: [PORT FROM ICS] Implement fast path for getParameters
BZ: 42090

Some applications call getParameters() directly from the callbacks they
have received from HAL. This causes deadlocks like the following:

A. HAL is flushing picture/video thread and message loop
   is blocked until the operation finishes
B. one of the pending picture/video messages, which was
   processed just before the flush, has called an app
   callback, which again calls HAL getParameters()
C. the app call to getParameters() is synchronous
D. deadlock results, as HAL/ControlThread is blocked on the
   flush call of step (A), and cannot process getParameters()

To protect HAL against this type of applications, modify getParameters
to be a nonblocking method, that can be run even if ControlThread is
blocked. This is implemented by maintaining a cache of flattened
parameters, which is updated whenever the active parameter set is
changed.

Change-Id: I50414914362a426ceaf7d331c7766913422c8183
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/54627
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/57639