libcamera2: [PORT FROM ICS] add doxygen mainpage
BZ: 47137

Add doc/libcamera-doc-mainpage.txt that contains the generic parts of
the component design documentation. This documentation gives an
introduction to the software, and helps readers to locate the most
important parts of the class documentation.

As an example, some sequence charts are included for continuous capture,
showing how also diagrams can be added via doxygen, and managed with git.

Change-Id: I2026143f5bbc6d4c6302d5a7862dbad1dfb3c641
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/55481
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/57912