libcamera2: [PORT FROM ICS] fix crash in case of errors during start preview
BZ: 45363

Problem: if any error happens during start preview the teardown
sequence crashes when tries to return the preview buffers to the
preview window. This happens if the Gfx buffers have been
unregistered by the camera service before the PreviewThread releases
them. This happens at different time because of the asynchronous
nature of the returnPreviewBuffers() call.
Solution: return the buffers to PreviewThread as soon as the error is
detected and make sure it completes. Since returnPreviewBuffers is async
call we need to call flushBuffers to make sure the command completes

Change-Id: Iae7e6b64f1a0909827310ed5566d995a1d2ccf78
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/55932
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/57673