libcamera2: [PORT FROM ICS] fix green screen in video mode (VGA)
BZ: 45363

Problem: Camera test application can set video mode with size
VGA. in this case the preview requested is 800x600 which has the same
AR as VGA. This is enough to let that preview size through and not
re-configuring the preview.
There is no check to clamp the preview resolution to the same size as
the video size in case the requested preview is bigger
This also happens with QVGA resolution

Solution: implement this check when processing the Static parameters
and the re-start preview.

Change-Id: I18cecebb519ac38a1c54f380a4d038a065e75306
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/55933
Reviewed-by: Alanen, Juha <juha.alanen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Hu Gang <gang.a.hu@intel.com>
Reviewed-on: http://android.intel.com:8080/57674