libcamera2: [PORT FROM ICS] HACK: Add load/unload FW to OLA HAL proxy
BZ: 47100

[2/6]
Part of the implementation of old style Acceleration
API into libcamera2. This is a temporary solution in order to
not disturb Social Camera.

In this change we add the methods in the HAL proxy  to load/unload
the firmware. We activate the OlaBufferService methods to
access those functions.
No functionality is executed yet in the ISP.

Change-Id: I08d6c857984db08a83209f9a651a27a5e970b994
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/56736
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/57916
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>