libcamera2: [PORT FROM ICS] HACK: Add Load/Unload Firmware to CtrlThread
BZ: 47100

[3/6]
Part of the implementation of old style Acceleration
API into libcamera2. This is a temporary solution in order to
not disturb Social Camera.

In this change we add the messages in the CtrlThread to handle
the request to load adn unload the firmware.

The actual IOCTL in the AtomISP class is left for next commit.

Change-Id: I2b3fda24feb1f847b5c8858a66574f6ccbc6e3cd
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/56737
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/57917
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>