libcamera2: [PORT FROM ICS] Fix race conditions when stopping camera
BZ: 46572

Problem: When stopping preview the preview buffers are returned
to the PreviewThread that returns them to the window.
Currently the CtrlThread sends an asynchronous message during
stopPreviewCore and continues

The camera service after preview is stop is frees the resources
from the window. If this happens before the PreviewThread has
time to return the buffers to the window it will crash

Solution:
Make the returnBuffers message from CtrlThread to PreviewThread
synchronous

Change-Id: I8622bea9035a2c47ad360f72029d5672bf0b50e2
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/56968
Reviewed-by: Vehmanen, Kai <kai.vehmanen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/57921
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>