libcamera2: flash settings stale when changing scene
BZ: 47896

When scene mode is changed, this may change the flash mode. The
processParamFlash() function, which acts on flash mode changes, was
however called after processParamSceneMode(), and thus flash settings
were not correctly updated after scene was changed.

Added a note to make sure the order of parameter processing is
maintained when doing future changes to this code.

Change-Id: I75b54952e606df6714a607897ca7d26067967457
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/58915
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>