libcamera2: enable HW jpeg encoding with libva
BZ: 47493

Issue:
We can't modify the libjpeg to realize the HW jpeg encoding.
So we must call our libva which is the HW jpeg encoder from
the camera HAL directly.

Solution:
One class of WrapperLibVA was added to the camera HAL.
It is used to enable the calling of libva.

This patch is the first step for enabling HW jpeg encoding with libva.
For the next, we will do the following steps.
1. add the image quality setting.
2. use libjpeg to instead of libskia.
3. optimize the performance.
4. consider the situation of using WIDI.

Change-Id: Ifcc03de58fdfe7421988d9b951c4430c3828997c
Signed-off-by: lyang56 <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/59165
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>