libcamera2: do not modify XNR while capture is running
BZ: 48284

If XNR setting was changed (either directly or via scene-mode change to
night mode), while ISP was still running the capture binary, a full
system freeze could be triggered.

To fix this issue, modify AtomISP to only allow changes in XNR state to
kernel when starting up capture. This follows the same pattern as already
used for ANR/low-light.

Change-Id: Ibd0a8888abaf2a5f1ea384c9d4950507fd3b8480
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/60194
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>