libcamera2: handle store metadata in still preview mode
BZ: 50769

Problem: CTS cases fail when the store metadata HAL call is
used in still preview mode

Root cause:
Original design did not allow for this. Once it was allowed another
problem surfaced. Metadata buffers cannot be allocated if not
in video mode, because this is the only state where the recording
buffers are already allocated.

Solution:
allow the HAL method to be used in still preview
and only allocate the metadata buffers when in ISP video mode

Change-Id: I7d46ec3a43275ca37ec44be4c422e4401d5df20a
Reviewed-on: http://android.intel.com:8080/60718
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>